soc/google/cpif: set the reinit_completion before the set_interrupt
A situation in which the usage count of netdev is not cleared occurs.
A timing issue about the wait_for_completion occurred
in the process of setting clat info to cp and receiving ack.
The complete_all is reinited before the wait_for_completion.
And, added defense code for situations that cp does not send ack.

Change-Id: I435e8be50e3829f471ebe0c002079e01c2764934
Signed-off-by: Namkyu Kim <namkyu78.kim@samsung.com>