soc: google: cpif: Switch to new OF thermal API
The thermal OF code has a new API allowing to migrate the OF
initialization to a simpler approach. The ops are no longer device
tree specific and are the generic ones provided by the core code.

Convert the ops to the thermal_zone_device_ops format and use the new
API to register the thermal zone with these generic ops.

Bug: 251537418
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I3a1e44d8306895959b18296cc0992ea67df73d84